add checkout inside add product option
